LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/align-items-baseline-column-vert-lr-grid-item-expected.txt

two
lines
hello

PASS #target > div 1
FAIL #target > div 2 assert_equals:
<div data-offset-x="56">hello</div>
offsetLeft expected 56 but got 57
